Move test to other issue integration tests
